requirements.txt
Flask==3.0.1
gunicorn==22.0.0
langchain==0.1.17
langchain_openai==0.1.4
langchain_experimental==0.0.57
langchain_community==0.0.36
weaviate-client==4.4.4
neo4j==5.20.0